[Serialization] Extract property and type checks into their own file
To split FirSerializationPluginClassChecker even further.

^KT-73207
